Removes talkspace link.
At the time it was the first list of hotline numbers I could find.
However, its a corporation acting as a sort of "DoorDash for
Therapists", and I'm not going to support that.

I should've done this years ago, but only got around to it now after
seeing FTC complaints against similar companies for selling sensitive
health information to advertisers.